[Decomp][ScriptMgr] Cohesion split: extract ScriptAction/DB-script loaders/binding to ScriptMgr*.cpp
Move three cohesive blocks out of the 3,702-line ScriptMgr.cpp into dedicated
TUs (verbatim move, no behaviour change; ScriptMgr.h unchanged, CMake
file(GLOB WorldHandlers/*.cpp) auto-registers them):

  ScriptMgrDbScripts.cpp  db_scripts table loaders (LoadScripts/LoadDbScripts/
                          LoadDbScriptStrings/CheckScriptTexts) + the file-local
                          GetSpellStartDBScriptPriority helper (def + all callers)
  ScriptAction.cpp        the ScriptAction runtime command executor
                          (HandleScriptStep + its object/target helpers)
  ScriptMgrBinding.cpp    script-name/binding registry + library-version query

ScriptMgr.cpp 3,702 -> 1,046. The On* scripting-hook dispatchers and library
load/unload stay in ScriptMgr.cpp for now (separate follow-up).
